moving main out of UI
