+++ android/app/src/main/AndroidManifest.xml
@@ -76,14 +76,14 @@
             android:value="AIzaSyAg2GgqVtmCLI6Ge73OdoU2xTYtIW_0Fp0"/>
 
 <!--        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~   Facebook configurations    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~       -->
-        <meta-data android:name="com.facebook.sdk.ApplicationId"
-            android:value="@string/facebook_app_id"/>
+        <!--<meta-data android:name="com.facebook.sdk.ApplicationId"
+            android:value="@string/facebook_app_id"/>-->
 
-        <activity android:name="com.facebook.FacebookActivity"
+        <!--<activity android:name="com.facebook.FacebookActivity"
             android:configChanges=
                 "keyboard|keyboardHidden|screenLayout|screenSize|orientation"
-            android:label="@string/app_name" />
-        <activity
+            android:label="@string/app_name" />-->
+        <!--<activity
             android:name="com.facebook.CustomTabActivity"
             android:exported="true">
             <intent-filter>
@@ -92,7 +92,7 @@
                 <category android:name="android.intent.category.BROWSABLE" />
                 <data android:scheme="@string/fb_login_protocol_scheme" />
             </intent-filter>
-        </activity>
+        </activity>-->
 
 
     </application>
